requirements.txt
Flask
python-dotenv
google-generativeai
pandas
matplotlib
Werkzeug
Pillow